split main.py into two files
